Check module enabled for timezones migration

+++ src/Database/Migrations/2020_07_07_055746_create_timezones_table.php
@@ -13,11 +13,13 @@ class CreateTimezonesTable extends Migration
 	 */
 	public function up()
 	{
-		Schema::create(config('world.migrations.timezones.table_name'), function (Blueprint $table) {
-			$table->id();
-			$table->foreignId('country_id');
-			$table->string('name');
-		});
+		if(config('world.modules.timezones')){
+			Schema::create(config('world.migrations.timezones.table_name'), function (Blueprint $table) {
+				$table->id();
+				$table->foreignId('country_id');
+				$table->string('name');
+			});
+		}
 	}
 
 	/**
@@ -27,6 +29,8 @@ class CreateTimezonesTable extends Migration
 	 */
 	public function down()
 	{
-		Schema::dropIfExists(config('world.migrations.timezones.table_name'));
+		if(config('world.modules.timezones')){
+			Schema::dropIfExists(config('world.migrations.timezones.table_name'));
+		}
 	}
 }
